Made another README for funsies
Added extra things because I realized I never actually made a new .csv after I fixed the merge conflicts. Anyhow, I did that and started a README for the new variables we're making, since it'll be real confusing once we start doing things and get confused. Feel free to update it, or whatever